+++ TruePath/Comparers/StrictStringPathComparer.cs
@@ -4,13 +4,13 @@
 
 namespace TruePath.Comparers;
 
-/// <summary>A strict comparer for comparing file paths using ordinal comparison.</summary>
-public class StrictPathComparer : IComparer<string>
+/// <summary>A strict comparer for comparing file paths using ordinal, case-sensitive comparison.</summary>
+public class StrictStringPathComparer : IComparer<string>
 {
     /// <summary>
-    /// Gets the singleton instance of the <see cref="StrictPathComparer"/> class.
+    /// Gets the singleton instance of the <see cref="StrictStringPathComparer"/> class.
     /// </summary>
-    public static readonly StrictPathComparer Instance = new();
+    public static readonly StrictStringPathComparer Instance = new();
 
     /// <inheritdoc cref="IComparer{T}.Compare"/>
     public int Compare(string? x, string? y)