+++ CHANGELOG.md
@@ -10,7 +10,7 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased] (1.6.0)
+## [1.6.0] - 2024-10-05
 ### Changed
 - Paths are now compared using a platform-specific comparer by default:
     - case-sensitive on Linux,
@@ -137,4 +137,5 @@ This is the first published version of the package. It doesn't contain any featu
 [1.3.0]: https://github.com/ForNeVeR/TruePath/compare/v1.2.1...v1.3.0
 [1.4.0]: https://github.com/ForNeVeR/TruePath/compare/v1.3.0...v1.4.0
 [1.5.0]: https://github.com/ForNeVeR/TruePath/compare/v1.4.0...v1.5.0
-[Unreleased]: https://github.com/ForNeVeR/TruePath/compare/v1.5.0...HEAD
+[1.5.0]: https://github.com/ForNeVeR/TruePath/compare/v1.5.0...v1.6.0
+[Unreleased]: https://github.com/ForNeVeR/TruePath/compare/v1.6.0...HEAD